test(gateway): extract auth tests
